Android fixes instead of compilers
Changed the name of the title in the Do It Station Activity
Fixed an issue where the LogInFragment was showing up twice.
Fixed an issue where RegistrationFragment would throw an error upon
creation.
-It was trying to encode null strings as JSON objects. Encoding as JSON
now only occurs if the text fields pass a series of tests
Changed some logic in Registration Fragment